requirements.txt
Django>=5.2,<5.3
Pillow>=11.0
pytesseract>=0.3.10
google-cloud-vision>=3.10.0
django-storages>=1.14.4
boto3>=1.35.0
firebase-admin>=6.5.0
python-dotenv>=1.0.1
gunicorn>=23.0.0
whitenoise>=6.8.2
dj-database-url>=2.2.0
psycopg[binary]>=3.2.1
